# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Property Lo API — Nigerian real estate listing platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface UnreadCountResponse {
  count: number;
}

export interface ChatIdResponse {
  id: string;
}

export interface User {
  id: string;
  email: string;
  full_name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  status: string;
  /** @nullable */
  avatar_base64?: string | null;
  created_at: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface SignupInput {
  email: string;
  password: string;
  full_name: string;
  phone?: string;
}

export interface VerifyOtpInput {
  email: string;
  otp: string;
}

export interface SigninInput {
  email: string;
  password: string;
}

export interface EmailInput {
  email: string;
}

export interface ResetPasswordInput {
  email: string;
  otp: string;
  new_password: string;
}

export interface ProfileUpdate {
  full_name: string;
  phone?: string;
}

export interface PasswordUpdate {
  current_password: string;
  new_password: string;
}

export interface AvatarUpdate {
  avatar_base64: string;
}

export interface Listing {
  id: string;
  title: string;
  description: string;
  type: string;
  purpose: string;
  price: number;
  location: string;
  state: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  images: string[];
  status: string;
  /** @nullable */
  contact_email?: string | null;
  /** @nullable */
  contact_phone?: string | null;
  /** @nullable */
  user_id?: string | null;
  created_at: string;
}

export interface ListingsResponse {
  listings: Listing[];
}

export interface ListingStats {
  active: number;
  houses: number;
  apartments: number;
  lands: number;
  locations: string[];
}

export interface ListingInput {
  title: string;
  description: string;
  type: string;
  purpose: string;
  price: number;
  location: string;
  state: string;
  size?: string;
  bedrooms?: number;
  bathrooms?: number;
  images?: string[];
  status?: string;
  contact_email?: string;
  contact_phone?: string;
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  type?: string;
  purpose?: string;
  price?: number;
  location?: string;
  state?: string;
  size?: string;
  bedrooms?: number;
  bathrooms?: number;
  images?: string[];
  status?: string;
  contact_email?: string;
  contact_phone?: string;
}

export interface FavoriteInput {
  listing_id: string;
}

export interface Chat {
  id: string;
  listing_id: string;
  listing_title: string;
  label: string;
  /** @nullable */
  last_message?: string | null;
  user_unread: number;
  admin_unread: number;
  updated_at: string;
  created_at: string;
}

export interface ChatMessage {
  id: string;
  content: string;
  is_admin: boolean;
  created_at: string;
}

export interface CreateChatInput {
  listing_id: string;
}

export interface MessageInput {
  content: string;
}

export interface SupportMessage {
  id: string;
  content: string;
  is_from_user: boolean;
  created_at: string;
}

export interface SupportChatResponse {
  messages: SupportMessage[];
}

export interface Ad {
  id: string;
  title: string;
  image_url: string;
  /** @nullable */
  redirect_url?: string | null;
  is_active: boolean;
  slide_order: number;
}

export interface AdInput {
  title: string;
  image_base64: string;
  redirect_url?: string;
  slide_order?: number;
  is_active?: boolean;
}

export interface AdUpdate {
  title?: string;
  image_base64?: string;
  redirect_url?: string;
  slide_order?: number;
  is_active?: boolean;
}

export interface ContactInput {
  first_name: string;
  last_name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface ContactMessage {
  id: string;
  first_name: string;
  last_name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  subject?: string | null;
  message: string;
  created_at: string;
}

export interface AdminStats {
  total_users: number;
  total_listings: number;
  total_chats: number;
  total_deals: number;
  new_users_this_month: number;
  new_listings_this_month: number;
  recent_users: User[];
  recent_listings: Listing[];
}

export type AdminAnalyticsChatLabelBreakdownItem = {
  label?: string;
  count?: number;
};

export type AdminAnalyticsLocationsWithMostListingsItem = {
  location?: string;
  count?: number;
};

export type AdminAnalyticsMostChattedHousesSaleItem = {
  title?: string;
  chat_count?: number;
};

export type AdminAnalyticsMostChattedHousesRentItem = {
  title?: string;
  chat_count?: number;
};

export type AdminAnalyticsMostChattedLandsItem = {
  title?: string;
  chat_count?: number;
};

export interface AdminAnalytics {
  chat_label_breakdown: AdminAnalyticsChatLabelBreakdownItem[];
  locations_with_most_listings: AdminAnalyticsLocationsWithMostListingsItem[];
  most_chatted_houses_sale: AdminAnalyticsMostChattedHousesSaleItem[];
  most_chatted_houses_rent: AdminAnalyticsMostChattedHousesRentItem[];
  most_chatted_lands: AdminAnalyticsMostChattedLandsItem[];
}

export interface AdminChat {
  id: string;
  listing_id: string;
  listing_title: string;
  label: string;
  /** @nullable */
  last_message?: string | null;
  user_unread: number;
  admin_unread: number;
  updated_at: string;
  /** @nullable */
  user_name?: string | null;
}

export interface AdminSupportChat {
  id: string;
  /** @nullable */
  user_name?: string | null;
  /** @nullable */
  last_message?: string | null;
  user_unread: number;
  admin_unread: number;
  updated_at: string;
}

export interface UserStatusUpdate {
  status: string;
}

export interface LabelUpdate {
  label: string;
}

export type GetListingsParams = {
type?: string;
purpose?: string;
location?: string;
search?: string;
status?: string;
limit?: string;
};

export type GetAdminListingsParams = {
type?: string;
status?: string;
page?: string;
limit?: string;
};

export type GetAdminUsersParams = {
search?: string;
status?: string;
page?: string;
limit?: string;
};

export type GetAdminChatsParams = {
label?: string;
page?: string;
limit?: string;
};

export type GetAdminSupportChatsParams = {
page?: string;
limit?: string;
};

export type GetContactMessagesParams = {
page?: string;
limit?: string;
};