# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Property Lo API — Nigerian real estate listing platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Account created and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: OTP resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/signin:
    post:
      operationId: signin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/admin/signin:
    post:
      operationId: adminSignin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninInput"
      responses:
        "200":
          description: Admin signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Reset code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/password:
    patch:
      operationId: updatePassword
      tags: [auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdate"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/avatar:
    patch:
      operationId: updateAvatar
      tags: [auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarUpdate"
      responses:
        "200":
          description: Avatar updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Listings ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    User:
      type: object
      required: [id, email, full_name, role, status, created_at]
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        avatar_base64:
          type: ["string", "null"]
        created_at:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    SignupInput:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
        password:
          type: string
        full_name:
          type: string
        phone:
          type: string
    VerifyOtpInput:
      type: object
      required: [email, otp]
      properties:
        email:
          type: string
        otp:
          type: string
    SigninInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [email, otp, new_password]
      properties:
        email:
          type: string
        otp:
          type: string
        new_password:
          type: string
    ProfileUpdate:
      type: object
      required: [full_name]
      properties:
        full_name:
          type: string
        phone:
          type: string
    PasswordUpdate:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
    AvatarUpdate:
      type: object
      required: [avatar_base64]
      properties:
        avatar_base64:
          type: string

    # ── Listing schemas ───────────────────────────────────────────────────────
...